sno-setup: use `rhel9-unknown` OS variant for guest VM
The libvirt stack really wants us to set this because it determines
various defaults. Since these scripts can be used for any version, we
can't just fixate on one (unless we start inspecting the RHCOS content
in the payload, which we tototally could). But for now at least, use
`rhel9-unknown` as a lower bound.