Drop `.sh` extension from executables
I want to build a layered FCOS image with all the requirements to run
these scripts built in (i.e. libvirt). In there, I want to have the
setup and cleanup scripts be in `/usr/bin` proper. So then it seems more
appropriate to not have the `.sh` extension.

I could do this extension change in the Containerfile only, but wanted
to see if there was interest in changing the canonical in-repo versions.