Add Containerfile to set up FCOS-based environment
For testing things like this, rather than installing libvirt locally, I
like to use throwaway VMs (and yes, use nested virt).

This Containerfile builds a layered FCOS image that ships all the
dependencies required in the image as well as the scripts. This image
can be built, and then a bootc/rpm-ostree-based OS can just rebase to it
to get started.

Obviously, it's not very hard to set up libvirt + git clone the repo,
but having an image makes it trivial to automate.